mockDatabase-build
Built a mock database. This was to understand formatting and how it will fit into the calendar and employee data warehouse. After practice and working with it I realize I will not be able to work on this further without the backend database. I need the backend in order to implement this properly.